+++ tests/unittests/gems_craft/lib_parsing/test_taxonomy_check.py
@@ -12,7 +12,6 @@
 
 import io
 from pathlib import Path
-from typing import Optional
 
 import pytest
 
@@ -21,8 +20,11 @@ from gems_craft.model.taxonomy import (
     Taxonomy,
     TaxonomyCategory,
     TaxonomyItem,
-    check_library_against_taxonomy,
     load_taxonomy,
+)
+from gems_craft.model.validation import (
+    check_library_against_taxonomy,
+    validate_libraries_against_taxonomy,
 )
 
 
@@ -34,8 +36,8 @@ def _make_category(cat_id: str, port_ids: list[str]) -> TaxonomyCategory:
     return TaxonomyCategory(id=cat_id, ports=[TaxonomyItem(id=p) for p in port_ids])
 
 
-def _parse_lib(yaml_content: str, taxonomy: Optional[Taxonomy] = None):
-    return parse_yaml_library(io.StringIO(yaml_content), taxonomy=taxonomy)
+def _parse_lib(yaml_content: str):
+    return parse_yaml_library(io.StringIO(yaml_content))
 
 
 # --- valid cases ---
@@ -421,7 +423,7 @@ library:
     check_library_against_taxonomy(lib, taxonomy)  # must not raise
 
 
-# --- parse_yaml_library wiring ---
+# --- validate_libraries_against_taxonomy ---
 
 _CONFORMING_LIB = """
 library:
@@ -458,34 +460,42 @@ library:
 """
 
 
-def test_parse_library_declaring_taxonomy_is_checked() -> None:
+def test_library_declaring_taxonomy_is_checked() -> None:
     taxonomy = _make_taxonomy(_make_category("production", ["injection_port"]))
-    lib = _parse_lib(_CONFORMING_LIB, taxonomy)
+    lib = _parse_lib(_CONFORMING_LIB)
+    validate_libraries_against_taxonomy([lib], taxonomy)  # must not raise
     assert lib.taxonomy == "test_taxonomy"
 
 
-def test_parse_library_declaring_taxonomy_raises_on_violation() -> None:
+def test_library_declaring_taxonomy_raises_on_violation() -> None:
     taxonomy = _make_taxonomy(_make_category("production", ["injection_port"]))
     with pytest.raises(ValueError, match="injection_port"):
-        _parse_lib(_VIOLATING_LIB, taxonomy)
+        validate_libraries_against_taxonomy([_parse_lib(_VIOLATING_LIB)], taxonomy)
 
 
-def test_parse_library_declaring_taxonomy_without_argument_raises() -> None:
+def test_library_declaring_taxonomy_without_argument_raises() -> None:
     with pytest.raises(ValueError, match="no taxonomy was provided"):
-        _parse_lib(_CONFORMING_LIB)
+        validate_libraries_against_taxonomy([_parse_lib(_CONFORMING_LIB)], None)
 
 
-def test_parse_library_with_mismatched_taxonomy_id_raises() -> None:
+def test_library_with_mismatched_taxonomy_id_raises() -> None:
     taxonomy = Taxonomy(
         id="other_taxonomy",
         categories=[_make_category("production", ["injection_port"])],
     )
     with pytest.raises(ValueError, match="other_taxonomy"):
-        _parse_lib(_CONFORMING_LIB, taxonomy)
+        validate_libraries_against_taxonomy([_parse_lib(_CONFORMING_LIB)], taxonomy)
 
 
-def test_parse_library_without_declared_taxonomy_is_not_checked() -> None:
+def test_library_without_declared_taxonomy_is_not_checked() -> None:
     """A model may carry a taxonomy-category without the library opting in."""
     taxonomy = _make_taxonomy(_make_category("production", ["injection_port"]))
-    lib = _parse_lib(_LIB_WITHOUT_TAXONOMY, taxonomy)
+    lib = _parse_lib(_LIB_WITHOUT_TAXONOMY)
+    validate_libraries_against_taxonomy([lib], taxonomy)  # must not raise
     assert lib.taxonomy is None
+
+
+def test_parsing_a_declaring_library_does_not_validate_it() -> None:
+    """Parsing is pure reading: conformance is the caller's business."""
+    lib = _parse_lib(_VIOLATING_LIB)  # must not raise
+    assert lib.taxonomy == "test_taxonomy"
